+++ Sources/MapboxDirections/RouteLeg.swift
@@ -25,6 +25,7 @@ open class RouteLeg: Codable {
         case expectedSegmentTravelTimes = "duration"
         case segmentSpeeds = "speed"
         case segmentCongestionLevels = "congestion"
+        case segmentMaximumSpeedLimits = "maxspeed"
     }
     
     // MARK: Creating a Leg
@@ -78,6 +79,12 @@ open class RouteLeg: Codable {
         expectedSegmentTravelTimes = try annotation?.decodeIfPresent([TimeInterval].self, forKey: .expectedSegmentTravelTimes)
         segmentSpeeds = try annotation?.decodeIfPresent([CLLocationSpeed].self, forKey: .segmentSpeeds)
         segmentCongestionLevels = try annotation?.decodeIfPresent([CongestionLevel].self, forKey: .segmentCongestionLevels)
+        
+        if let speedLimitDescriptors = try annotation?.decodeIfPresent([SpeedLimitDescriptor].self, forKey: .segmentMaximumSpeedLimits) {
+            segmentMaximumSpeedLimits = speedLimitDescriptors.map { Measurement<UnitSpeed>(speedLimitDescriptor: $0) }
+        } else {
+            segmentMaximumSpeedLimits = nil
+        }
     }
     
     public func encode(to encoder: Encoder) throws {
@@ -90,12 +97,16 @@ open class RouteLeg: Codable {
         try container.encode(expectedTravelTime, forKey: .expectedTravelTime)
         try container.encode(profileIdentifier, forKey: .profileIdentifier)
         
-        if segmentDistances != nil || expectedSegmentTravelTimes != nil || segmentSpeeds != nil || segmentCongestionLevels != nil {
+        if segmentDistances != nil || expectedSegmentTravelTimes != nil || segmentSpeeds != nil || segmentCongestionLevels != nil || segmentMaximumSpeedLimits != nil {
             var annotationContainer = container.nestedContainer(keyedBy: AnnotationCodingKeys.self, forKey: .annotation)
             try annotationContainer.encodeIfPresent(segmentDistances, forKey: .segmentDistances)
             try annotationContainer.encodeIfPresent(expectedSegmentTravelTimes, forKey: .expectedSegmentTravelTimes)
             try annotationContainer.encodeIfPresent(segmentSpeeds, forKey: .segmentSpeeds)
             try annotationContainer.encodeIfPresent(segmentCongestionLevels, forKey: .segmentCongestionLevels)
+            
+            if let speedLimitDescriptors = segmentMaximumSpeedLimits?.map({ SpeedLimitDescriptor(speed: $0) }) {
+                try annotationContainer.encode(speedLimitDescriptors, forKey: .segmentMaximumSpeedLimits)
+            }
         }
     }
 
@@ -126,16 +137,39 @@ open class RouteLeg: Codable {
 
      Each route step object corresponds to a distinct maneuver and the approach to the next maneuver.
 
-     This array is empty if the `includesSteps` property of the original `RouteOptions` object is set to `false`.
+     This array is empty if the original `RouteOptions` object’s `RouteOptions.includesSteps` property is set to `false`.
      */
     public let steps: [RouteStep]
+    
+    /**
+     The ranges of each step’s segments within the overall leg.
+     
+     Each range corresponds to an element of the `steps` property. Use this property to safely subscript segment-based properties such as `segmentCongestionLevels` and `segmentMaximumSpeedLimits`.
+     
+     This array is empty if the original `RouteOptions` object’s `RouteOptions.includesSteps` property is set to `false`.
+     */
+    public private(set) lazy var segmentRangesByStep: [Range<Int>] = {
+        var segmentRangesByStep: [Range<Int>] = []
+        var currentStepStartIndex = 0
+        for step in steps {
+            if let coordinates = step.shape?.coordinates {
+                let stepCoordinateCount = step.maneuverType == .arrive ? coordinates.count : coordinates.dropLast().count
+                let currentStepEndIndex = currentStepStartIndex.advanced(by: stepCoordinateCount)
+                segmentRangesByStep.append(currentStepStartIndex..<currentStepEndIndex)
+                currentStepStartIndex = currentStepEndIndex
+            } else {
+                segmentRangesByStep.append(currentStepStartIndex..<currentStepStartIndex)
+            }
+        }
+        return segmentRangesByStep
+    }()
     
     // MARK: Getting Per-Segment Attributes Along the Leg
     
     /**
      An array containing the distance (measured in meters) between each coordinate in the route leg geometry.
 
-     This property is set if the `RouteOptions.attributeOptions` property contains `.distance`.
+     This property is set if the `RouteOptions.attributeOptions` property contains `AttributeOptions.distance`.
      */
     open var segmentDistances: [CLLocationDistance]?
 
@@ -144,7 +178,7 @@ open class RouteLeg: Codable {
 
      These values are dynamic, accounting for any conditions that may change along a segment, such as traffic congestion if the profile identifier is set to `.automobileAvoidingTraffic`.
 
-     This property is set if the `RouteOptions.attributeOptions` property contains `.expectedTravelTime`.
+     This property is set if the `RouteOptions.attributeOptions` property contains `AttributeOptions.expectedTravelTime`.
      */
     open var expectedSegmentTravelTimes: [TimeInterval]?
 
@@ -153,7 +187,7 @@ open class RouteLeg: Codable {
 
      These values are dynamic; rather than speed limits, they account for the road’s classification and/or any traffic congestion (if the profile identifier is set to `.automobileAvoidingTraffic`).
 
-     This property is set if the `RouteOptions.attributeOptions` property contains `.speed`.
+     This property is set if the `RouteOptions.attributeOptions` property contains `AttributeOptions.speed`.
      */
     open var segmentSpeeds: [CLLocationSpeed]?
 
@@ -164,10 +198,21 @@ open class RouteLeg: Codable {
 
      You can color-code a route line according to the congestion level along each segment of the route.
 
-     This property is set if the `RouteOptions.attributeOptions` property contains `.congestionLevel`.
+     This property is set if the `RouteOptions.attributeOptions` property contains `AttributeOptions.congestionLevel`.
      */
     open var segmentCongestionLevels: [CongestionLevel]?
-
+    
+    /**
+     An array containing the maximum speed limit along each road segment along the route leg’s shape.
+     
+     The maximum speed may be an advisory speed limit for segments where legal limits are not posted, such as highway entrance and exit ramps. If the speed limit along a particular segment is unknown, it is represented in the array by a measurement whose value is negative. If the speed is unregulated along the segment, such as on the German _Autobahn_ system, it is represented by a measurement whose value is `Double.infinity`.
+     
+     Speed limit data is available in [a number of countries and territories worldwide](https://docs.mapbox.com/help/how-mapbox-works/directions/).
+     
+     This property is set if the `RouteOptions.attributeOptions` property contains `AttributeOptions.maximumSpeedLimit`.
+     */
+    open var segmentMaximumSpeedLimits: [Measurement<UnitSpeed>?]?
+    
     // MARK: Getting Statistics About the Leg
 
     /**
@@ -214,6 +259,7 @@ extension RouteLeg: Equatable {
             lhs.expectedSegmentTravelTimes == rhs.expectedSegmentTravelTimes &&
             lhs.segmentSpeeds == rhs.segmentSpeeds &&
             lhs.segmentCongestionLevels == rhs.segmentCongestionLevels &&
+            lhs.segmentMaximumSpeedLimits == rhs.segmentMaximumSpeedLimits &&
             lhs.name == rhs.name &&
             lhs.distance == rhs.distance &&
             lhs.expectedTravelTime == rhs.expectedTravelTime &&