+++ Sources/MapboxDirections/RouteStep.swift
@@ -244,6 +244,27 @@ public enum ManeuverDirection: String, Codable {
     case uTurn = "uturn"
 }
 
+/**
+ A road sign design standard.
+ 
+ A sign standard can affect how a user interface should display information related to the road. For example, a speed limit from the `RouteLeg.segmentMaximumSpeedLimits` property may appear in a different-looking view depending on the `RouteStep.speedLimitSign` property.
+ */
+public enum SignStandard: String, Codable {
+    /**
+     The [Manual on Uniform Traffic Control Devices](https://en.wikipedia.org/wiki/Manual_on_Uniform_Traffic_Control_Devices).
+     
+     This standard has been adopted by the United States and Canada, and several other countries have adopted parts of the standard as well.
+     */
+    case mutcd
+    
+    /**
+     The [Vienna Convention on Road Signs and Signals](https://en.wikipedia.org/wiki/Vienna_Convention_on_Road_Signs_and_Signals).
+     
+     This standard is prevalent in Europe and parts of Asia and Latin America. Countries in southern Africa and Central America have adopted similar regional standards.
+     */
+    case viennaConvention
+}
+
 extension String {
     internal func tagValues(separatedBy separator: String) -> [String] {
         return components(separatedBy: separator).map { $0.trimmingCharacters(in: .whitespaces) }.filter { !$0.isEmpty }
@@ -370,6 +391,8 @@ open class RouteStep: Codable {
         case maneuver
         case pronunciation
         case rotaryPronunciation = "rotary_pronunciation"
+        case speedLimitSignStandard = "speedLimitSign"
+        case speedLimitUnit
         case transportType = "mode"
     }
     
@@ -406,10 +429,12 @@ open class RouteStep: Codable {
      - parameter destinationCodes: Any route reference codes that appear on guide signage for the road leading from this step’s maneuver to the next step’s maneuver.
      - parameter destinations: Destinations, such as [control cities](https://en.wikipedia.org/wiki/Control_city), that appear on guide signage for the road leading from this step’s maneuver to the next step’s maneuver.
      - parameter intersections: An array of intersections along the step.
+     - parameter speedLimitSignStandard: The sign design standard used for speed limit signs along the step.
+     - parameter speedLimitUnit: The unit of speed limits on speed limit signs along the step.
      - parameter instructionsSpokenAlongStep: Instructions about the next step’s maneuver, optimized for speech synthesis.
      - parameter instructionsDisplayedAlongStep: Instructions about the next step’s maneuver, optimized for display in real time.
      */
-    public init(transportType: TransportType, maneuverLocation: CLLocationCoordinate2D, maneuverType: ManeuverType, maneuverDirection: ManeuverDirection? = nil, instructions: String, initialHeading: CLLocationDirection? = nil, finalHeading: CLLocationDirection? = nil, drivingSide: DrivingSide, exitCodes: [String]? = nil, exitNames: [String]? = nil, phoneticExitNames: [String]? = nil, distance: CLLocationDistance, expectedTravelTime: TimeInterval, names: [String]? = nil, phoneticNames: [String]? = nil, codes: [String]? = nil, destinationCodes: [String]? = nil, destinations: [String]? = nil, intersections: [Intersection]? = nil, instructionsSpokenAlongStep: [SpokenInstruction]? = nil, instructionsDisplayedAlongStep: [VisualInstructionBanner]? = nil) {
+    public init(transportType: TransportType, maneuverLocation: CLLocationCoordinate2D, maneuverType: ManeuverType, maneuverDirection: ManeuverDirection? = nil, instructions: String, initialHeading: CLLocationDirection? = nil, finalHeading: CLLocationDirection? = nil, drivingSide: DrivingSide, exitCodes: [String]? = nil, exitNames: [String]? = nil, phoneticExitNames: [String]? = nil, distance: CLLocationDistance, expectedTravelTime: TimeInterval, names: [String]? = nil, phoneticNames: [String]? = nil, codes: [String]? = nil, destinationCodes: [String]? = nil, destinations: [String]? = nil, intersections: [Intersection]? = nil, speedLimitSignStandard: SignStandard? = nil, speedLimitUnit: UnitSpeed? = nil, instructionsSpokenAlongStep: [SpokenInstruction]? = nil, instructionsDisplayedAlongStep: [VisualInstructionBanner]? = nil) {
         self.transportType = transportType
         self.maneuverLocation = maneuverLocation
         self.maneuverType = maneuverType
@@ -429,6 +454,8 @@ open class RouteStep: Codable {
         self.destinationCodes = destinationCodes
         self.destinations = destinations
         self.intersections = nil
+        self.speedLimitSignStandard = speedLimitSignStandard
+        self.speedLimitUnit = speedLimitUnit
         self.instructionsSpokenAlongStep = nil
         self.instructionsDisplayedAlongStep = nil
     }
@@ -473,6 +500,12 @@ open class RouteStep: Codable {
         try maneuver.encodeIfPresent(maneuverLocation, forKey: .location)
         try maneuver.encodeIfPresent(initialHeading, forKey: .initialHeading)
         try maneuver.encodeIfPresent(finalHeading, forKey: .finalHeading)
+        
+        try container.encodeIfPresent(speedLimitSignStandard, forKey: .speedLimitSignStandard)
+        if let speedLimitUnit = speedLimitUnit,
+            let unit = SpeedLimitDescriptor.UnitDescriptor(unit: speedLimitUnit) {
+            try container.encode(unit, forKey: .speedLimitUnit)
+        }
     }
     
     public required init(from decoder: Decoder) throws {
@@ -522,6 +555,9 @@ open class RouteStep: Codable {
         exitCodes = road.exitCodes
         destinations = road.destinations
         destinationCodes = road.destinationCodes
+        
+        speedLimitSignStandard = try container.decodeIfPresent(SignStandard.self, forKey: .speedLimitSignStandard)
+        speedLimitUnit = (try container.decodeIfPresent(SpeedLimitDescriptor.UnitDescriptor.self, forKey: .speedLimitUnit))?.describedUnit
         
         let type = maneuverType
         if type == .takeRotary || type == .takeRoundabout {
@@ -704,6 +740,20 @@ open class RouteStep: Codable {
     */
     public let intersections: [Intersection]?
     
+    /**
+     The sign design standard used for speed limit signs along the step.
+     
+     This standard affects how corresponding speed limits in the `RouteLeg.segmentMaximumSpeedLimits` property should be displayed.
+     */
+    public let speedLimitSignStandard: SignStandard?
+    
+    /**
+     The unit of speed limits on speed limit signs along the step.
+     
+     This standard affects how corresponding speed limits in the `RouteLeg.segmentMaximumSpeedLimits` property should be displayed.
+     */
+    public let speedLimitUnit: UnitSpeed?
+    
     // MARK: Getting Details About the Next Maneuver
     
     /**
@@ -750,6 +800,9 @@ extension RouteStep: Equatable {
             lhs.codes == rhs.codes &&
             lhs.destinationCodes == rhs.destinationCodes &&
             lhs.destinations == rhs.destinations &&
+            
+            lhs.speedLimitSignStandard == rhs.speedLimitSignStandard &&
+            lhs.speedLimitUnit == rhs.speedLimitUnit &&
             
             lhs.intersections == rhs.intersections &&
             lhs.instructionsSpokenAlongStep == rhs.instructionsSpokenAlongStep &&